Simplified EquipmentSetItem
